requirements.txt
aiogram==3.7.0
google-genai
apscheduler
fastapi
uvicorn[standard]
python-dotenv
jinja2
aiofiles
